Fixed BybitTrade.Size

+++ bybit/types/markets/trades.go
@@ -7,7 +7,7 @@ type Trade struct {
 	TradeTimeMs   string    `json:"trade_time_ms"`
 	Symbol        string    `json:"symbol"`
 	Side          string    `json:"side"`
-	Size          float64   `json:"size"`
+	Size          float32   `json:"size"`
 	Price         string    `json:"price"`
 	TickDirection string    `json:"tick_direction"`
 	TradeID       string    `json:"trade_id"`
